Create fitz.i
This pertains to Page method `get_drawings()`. 
The set of keys for "fs" paths already is complete by definition - so only fill and stroke paths require special attention.
Each type should have the complete list of keys as a "fs" path would have, so we generate `None` values where necessary for "f" and "s" paths only.
In view of performance, this is probably faster then subclassing `dict` and giving it the `__missing__()` method to achieve the same thing.

Unrelated to this fix:
When dealing with Optional Content Properties, the array of `/Locked` optional content groups is now also supported for input and output.